requirements.txt
# Minimal requirements
Django>=1.9
https://static.dpaw.wa.gov.au/static/py/dpaw-utils/dist/dpaw-utils-0.3a4.tar.gz
wagtail>=1.3.1
wagtailplus>=0.3.7
django-queryset-csv>=0.3.3
django-extensions>=1.6.1
django-mptt>=0.8.0
django-leaflet>=0.18.0
django-reversion>=1.10.0
python-dateutil>=2.4.2
django-model-utils>=2.4
restless>=2.0.1
django-auth-ldap>=1.2.7
django-ipware>=1.1.2
python-social-auth>=0.2.13
graphviz>=0.4.8
json2html>=1.0.1
python-ldap==2.4.25
djangorestframework==3.3.2
django-filter==0.12.0
